move keystore functionalities to account pkg
