+++ account/account_test.go
@@ -12,7 +12,6 @@ import (
 	"github.com/joho/godotenv"
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/starknet.go/account"
-	"github.com/NethermindEth/starknet.go/curve"
 	"github.com/NethermindEth/starknet.go/mocks"
 	"github.com/NethermindEth/starknet.go/rpc"
 	"github.com/NethermindEth/starknet.go/test"
@@ -128,7 +127,7 @@ func TestTransactionHashInvoke(t *testing.T) {
 	for _, test := range testSet {
 
 		t.Run("Transaction hash", func(t *testing.T) {
-			ks := curve.NewMemKeystore()
+			ks := account.NewMemKeystore()
 			if test.SetKS {
 				privKeyBI, ok := new(big.Int).SetString(test.PrivKey.String(), 0)
 				require.True(t, ok)
@@ -219,7 +218,7 @@ func TestChainIdMOCK(t *testing.T) {
 
 	for _, test := range testSet {
 		mockRpcProvider.EXPECT().ChainID(context.Background()).Return(test.ChainID, nil)
-		account, err := account.NewAccount(mockRpcProvider, &felt.Zero, "pubkey", curve.NewMemKeystore())
+		account, err := account.NewAccount(mockRpcProvider, &felt.Zero, "pubkey", account.NewMemKeystore())
 		require.NoError(t, err)
 		require.Equal(t, account.ChainId.String(), test.ExpectedID)
 	}
@@ -250,7 +249,7 @@ func TestChainId(t *testing.T) {
 		require.NoError(t, err, "Error in rpc.NewClient")
 		provider := rpc.NewProvider(client)
 
-		account, err := account.NewAccount(provider, &felt.Zero, "pubkey", curve.NewMemKeystore())
+		account, err := account.NewAccount(provider, &felt.Zero, "pubkey", account.NewMemKeystore())
 		require.NoError(t, err)
 		require.Equal(t, account.ChainId.String(), test.ExpectedID)
 	}
@@ -291,7 +290,7 @@ func TestSignMOCK(t *testing.T) {
 	for _, test := range testSet {
 		privKeyBI, ok := new(big.Int).SetString(test.PrivKey.String(), 0)
 		require.True(t, ok)
-		ks := curve.NewMemKeystore()
+		ks := account.NewMemKeystore()
 		ks.Put(test.Address.String(), privKeyBI)
 
 		mockRpcProvider.EXPECT().ChainID(context.Background()).Return(test.ChainId, nil)
@@ -382,7 +381,7 @@ func TestAddInvoke(t *testing.T) {
 		provider := rpc.NewProvider(client)
 
 		// Set up ks
-		ks := curve.NewMemKeystore()
+		ks := account.NewMemKeystore()
 		if test.SetKS {
 			fakePrivKeyBI, ok := new(big.Int).SetString(test.PrivKey.String(), 0)
 			require.True(t, ok)
@@ -419,7 +418,7 @@ func TestAddDeployAccountDevnet(t *testing.T) {
 	fakeUserPub := utils.TestHexToFelt(t, fakeUser.PublicKey)
 
 	// Set up ks
-	ks := curve.NewMemKeystore()
+	ks := account.NewMemKeystore()
 	fakePrivKeyBI, ok := new(big.Int).SetString(fakeUser.PrivateKey, 0)
 	require.True(t, ok)
 	ks.Put(fakeUser.PublicKey, fakePrivKeyBI)
@@ -465,7 +464,7 @@ func TestTransactionHashDeployAccountTestnet(t *testing.T) {
 
 	ExpectedHash := utils.TestHexToFelt(t, "0x5b6b5927cd70ad7a80efdbe898244525871875c76540b239f6730118598b9cb")
 	ExpectedPrecomputeAddr := utils.TestHexToFelt(t, "0x88d0038623a89bf853c70ea68b1062ccf32b094d1d7e5f924cda8404dc73e1")
-	ks := curve.NewMemKeystore()
+	ks := account.NewMemKeystore()
 	fakePrivKeyBI, ok := new(big.Int).SetString(PrivKey.String(), 0)
 	require.True(t, ok)
 	ks.Put(PubKey.String(), fakePrivKeyBI)
@@ -509,7 +508,7 @@ func TestTransactionHashDeclare(t *testing.T) {
 	require.NoError(t, err, "Error in rpc.NewClient")
 	provider := rpc.NewProvider(client)
 
-	acnt, err := account.NewAccount(provider, &felt.Zero, "", curve.NewMemKeystore())
+	acnt, err := account.NewAccount(provider, &felt.Zero, "", account.NewMemKeystore())
 	require.NoError(t, err)
 
 	tx := rpc.DeclareTxnV2{
